requirementsa.txt
# Graph database driver
neo4j==5.26.0

# Embeddings (runs 100% locally — downloads model once ~90MB)
sentence-transformers==2.7.0
torch==2.4.1
transformers==4.45.0

# Data handling
pandas==2.2.2
numpy==1.26.4

# Web UI
streamlit==1.40.0
plotly==5.24.1

# REST API (bonus Part F)
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0

# Utilities
python-dotenv==1.0.1
scikit-learn==1.5.2

